Refactor project structure for better organization
Main program files are moved into a new directory named 'example'. This is done to make the structure of the project tidier and simpler to understand. It separates example implementation from the main code base, making it easier to distinguish between the part of the library and examples for using the library.