Update handling of user interaction and improve animation transitions
This commit replaces the detectTapGestures/onLongPress method with the clickable Modifier to intercept user inputs to initiate animations. This update follows the best practices of the Compose UI framework for handling user interaction in a simpler and more efficient manner. Additionally, it introduces 'previousRoute' variable in place of 'NavTransitions.lastRoute' for managing state during navigation transitions. This event will provide a more accurate and efficient way of managing the system and improve the overall performance of the navigation transitions.