Add README.md for Compose Nav Transitions
Introduced a README.md file for the Compose Nav Transitions library. The README includes a brief overview of the library, badges for quick library stats, a demo GIF and the MIT License text. This provides all necessary basic details for any developer who wants to understand or contribute to the library.