Add alpha control and visibility toggle to NavTransition
Added alpha control to enhance the transition effect & created a visibility state in the NavTransitionContainer. This change was made to smooth out the navigation transitions and to handle the visibility of the transition elements more efficiently. Removed redundant animation check. Adjusted background color of the animated element to light gray for better visibility.