Add contribution guide to README.md
A section on how to contribute to this project was added to README.md. This change aims to make contributing as easy and transparent as possible. It includes steps on how to submit bug reports, propose new features, become a maintainer and more. This addition endorses an open and inclusive culture for project contributions.