Set up publishing and signing for compose-nav-transition library
Modified the build.gradle.kts of the compose-nav-transition library to set up Maven publishing and the signing of the release version. This allows the release versions of the library to be published to 'https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/' and automatically get signed. It includes the use of a properties file for the OSSRH account credentials to ensure secure access. Furthermore, readme.md has been expanded to help users comprehend installation and usage.