Fix repetitive sharedElement calls in NavTransitionScope
A check has been added to see if a sharedElement with the given tag has already been handled before proceeding with the calculation of Rect and updating the shared elements. This change prevents redundant calls to the sharedElement function in NavTransitionScope class. This optimization reduces unnecessary computations, thereby improving efficiency.