Add new Element.kt file with abstract Element class
Added a new file Element.kt under the compose-nav-transition module. This class acts as an abstract base class for different UI elements(like TextElement, ImageElement, and ImageVectorElement) that can be displayed in the application. This change allows us to represent different types of elements more intuitively and improves the extension capabilities of our code. This also provides a uniform way to handle various UI elements.