Add element parameters to sharedElement function in NavTransitionScope
Elaborated the sharedElement function to include an optional 'element' parameter. This now allows an instance of Element to be associated with a tag, enabling us to not just associate an element's attributes with the tag, but the element itself as well. This widens the scope of the sharedElement function and streamlines the process of referring and recalling a UI element's attributes and its type.