Refactor sharedElement function in NavTransitionScope
Updated the logic in the sharedElement() function in NavTransitionScope, replacing 'rects' with 'elements' and adding an 'element' parameter to include an instance of Element, instead of just dealing with a element's attributes. This enhances the function in two ways: 1) it simplifies handling a UI element's attributes and its type. 2) It provides a more flexible way to reference and recall a UI element and its attributes. This change is the groundwork for future enhancements.