Update compose-nav-transitions version in README
Upgraded the version of compose-nav-transitions library in the README.md from "0.1.0-alpha01" to "0.2.0-alpha01". Also added a version guide table to track compatibility with the Navigation Compose library. The library was updated to incorporate new updates and enhancements. This practice aids in version control and effective handling of project dependencies.