Add alternate method in README.md for shared element navigation
New block of code is added to README.md to improve clarity. This provides an alternate way to use sharedElement in NavTransitionScope. It shows how to add NavTransitionScope to the MyScreen function and how to add sharedElement to text and image modifiers. This method can be beneficial for developers who prefer a different approach or find this way more understandable.