Add preview object in NavTransitionScope
The commit adds a preview object in the NavTransitionScope class. The necessity of this commit is to facilitate easier and quicker access for developers when they want to preview transition components during development, thus improving their efficiency and productivity. It may also be used for demo purposes or to test out shared elements before implementing them.