Add NotificationScreen in the Navigation Transition Example.
A new notification screen was added in the Navigation Transition Example project. This screen consists of a notification list that uses LazyColumn and simulates a typical scenario where a user sees a notifications list and can navigate back. The reason was to showcase the capability of Compose Navigation Transition library within an example project, demonstrating how transitions act in these situations. Additionally, dependencies like Icons and images were also added.