Update icon assets
Added a new SVG file for complex gradients and transition animations. Also updated the Android launcher icons to use new assets and changed the background reference from a drawable to a color resource, for increased performance and smoother rendering. This change enhances the visual aesthetics and performance of the app.