Update NavTransitionGraphBuilder to accept named navigation arguments
Navigation arguments support is added to NavTransitionGraphBuilder's composable function as per application requirements. This enhancement will allow each composable in the navigation graph to accept a list of NamedNavArguments which broadens the scope of data passed between the navigation graphs. This will help to ensure seamless navigation across composables with better state management.