Update Element class in compose-nav-transition module
Updated the Element class in the compose-nav-transition module. Added 'id' property to uniquely identify each instance of this class. Changed the properties from mutable to read-only. Removed 'alpha' property which was found redundant. Modified the order of properties to increase readability. This refactoring is done to improve element identification during navigational transitions in a Compose UI.