Replace UUID with DpRect in Element class
The Element class's ID field, which was previously a random UUID, has been replaced with a DpRect field called 'rect'. This was done to allow more precise and fluid transitions in the Navigation Compose module. It removes the need to generate a UUID for every Element instance, thus potentially improving performance.