Refactor NavTransitionHost for enhanced transitions
Modified NavTransitionHost to improve precision and fluidity of navigation transitions. The Element class's ID field, previously a random UUID, was replaced by a DpRect called 'rect'. This removes the need for generating UUID for every Element instance, with potential performance benefits. Changes also allow for a more effective reset and track of current and previous elements in the navigation scope.