Refactor Element subclasses to adopt a default Modifier
Shifted the declaration of the `modifier` variable in `Element` and its subclasses `TextElement`, `ImageElement`, and `IconElement` to allow a default value to be set for it. This simplifies the construction of these objects since the `modifier` now has a default value and need not be specified when calling the constructor, leading to cleaner and less error-prone code. This default value can be overridden if needed.