Refactor NavTransitionScope to improve efficiency
The NavTransitionScope has been amended to make use of MutableStateFlow for managing shared elements, allowing for more efficient tracking and updating of state changes. This revision removes the usage of mutableStateOf and alphaMap properties in favour of MutableStateFlow, hence streamlining and simplifying the state handling. It includes a new initializer for the class, taking in an optional list of Element objects, and a companion object that takes advantage of Kotlin's listSaver function for state restoration.