Refactor screen navigation with animations and enum type
Screen navigation refactored for a smoother user experience. An enum class `ScreenState` was added to manage screen states and to replace the previous mutable integer state. AnimatedVisibility Composable added in MainActivity along with fadeIn and fadeOut animations for screen transitions. Making the application transition between screens more visually pleasing with a fadeIn and fadeOut effect over a duration of 600 milliseconds.