Add navigation transition support in Element class
Element class is refactored from a data class to a class with an internal constructor to support navigation transitions between screens. A mutable state for the 'rect' property is introduced to allow dynamic changes. New functions 'connect' and 'transitionDpRect' are added for managing connections between elements and transitioning states respectively. An instance 'None' of the Element class is declared in the companion object to cater for null cases.